+++ CHANGELOG.md
@@ -4,6 +4,23 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com), and this project adheres to [Semantic Versioning](https://semver.org).
 
+## [Unreleased]
+### Added
+* *Nothing*
+
+### Changed
+* *Nothing*
+
+### Deprecated
+* *Nothing*
+
+### Removed
+* Removed deprecated stuff.
+
+### Fixed
+* *Nothing*
+
+
 ## [4.5.0] - 2022-08-05
 ### Added
 * Added new `MonologFactory` for monolog 3, and a couple new logger utilities.